+++ src/components/App.tsx
@@ -1,12 +1,14 @@
 import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Route, Routes, useLocation } from 'react-router-dom';
+import { Outlet, Route, Routes } from 'react-router-dom';
 
-import { saveUrlForRedirection } from '@graasp/sdk';
-import { Loader, withAuthorization } from '@graasp/ui';
+import { Box } from '@mui/material';
 
-import { DOMAIN } from '@/config/env';
-import { SIGN_IN_PATH } from '@/config/externalPaths';
+import { buildSignInPath, getCurrentAccountLang } from '@graasp/sdk';
+import { DEFAULT_LANG } from '@graasp/translations';
+import { Loader, SignedInWrapper } from '@graasp/ui';
+
+import { GRAASP_AUTH_HOST } from '@/config/env';
 import { hooks } from '@/config/queryClient';
 
 import {
@@ -19,7 +21,9 @@ import {
   USERS_ANALYTICS_PATH,
   buildItemPath,
 } from '../config/paths';
-import ContextsWrapper from './context/ContextsWrapper';
+import DataProvider from './context/DataProvider';
+import ViewDataProvider from './context/ViewDataProvider';
+import Navigator from './layout/Navigator';
 import PageWrapper from './layout/PageWrapper';
 import HomePageWrapper from './pages/HomePage';
 import AppsAnalyticPage from './pages/Item/AppsAnalyticPage';
@@ -31,78 +35,64 @@ import UsersAnalyticPage from './pages/Item/UsersAnalyticPage';
 import MyAnalyticsPage from './pages/MyAnalyticsPage';
 
 const App = (): JSX.Element => {
-  const { data: currentMember, isLoading } = hooks.useCurrentMember();
+  const { data: currentAccount, isLoading } = hooks.useCurrentMember();
   const { i18n } = useTranslation();
-  const { pathname } = useLocation();
 
   useEffect(() => {
-    if (currentMember?.extra?.lang !== i18n.language) {
-      i18n.changeLanguage(currentMember?.extra?.lang ?? 'en');
+    const lang = getCurrentAccountLang(currentAccount, DEFAULT_LANG);
+    if (lang) {
+      i18n.changeLanguage(lang);
     }
-  }, [currentMember]);
+  }, [currentAccount]);
 
   if (isLoading) {
     return <Loader />;
   }
-
-  const withAuthorizationProps = {
-    currentMember,
-    redirectionLink: SIGN_IN_PATH,
-    onRedirect: () => {
-      // save current url for later redirection after sign in
-      saveUrlForRedirection(pathname, DOMAIN);
-    },
-  };
-
-  const HomeWrapperWithAuth = withAuthorization(
-    HomePageWrapper,
-    withAuthorizationProps,
-  );
-  const ItemWrapperWithAuth = withAuthorization(
-    ItemPage,
-    withAuthorizationProps,
-  );
-  const MyAnalyticsWithAuth = withAuthorization(
-    MyAnalyticsPage,
-    withAuthorizationProps,
-  );
 
   return (
     <Routes>
       <Route path={EMBEDDED_ITEM_PATH} element={<ItemPage />} />
 
       <Route
-        path={HOME_PATH}
         element={
-          <PageWrapper>
-            <HomeWrapperWithAuth />
-          </PageWrapper>
-        }
-      />
-      <Route
-        path={MY_ANALYTICS_PATH}
-        element={
-          <PageWrapper>
-            <MyAnalyticsWithAuth />
-          </PageWrapper>
-        }
-      />
-
-      <Route
-        path={buildItemPath()}
-        element={
-          <ContextsWrapper>
-            <PageWrapper>
-              <ItemWrapperWithAuth />
-            </PageWrapper>
-          </ContextsWrapper>
+          <SignedInWrapper
+            currentAccount={currentAccount}
+            redirectionLink={buildSignInPath({
+              host: GRAASP_AUTH_HOST,
+              redirectionUrl: window.location.toString(),
+            })}
+          >
+            <DataProvider>
+              <PageWrapper>
+                <Outlet />
+              </PageWrapper>
+            </DataProvider>
+          </SignedInWrapper>
         }
       >
-        <Route index element={<GeneralAnalyticsPage />} />
-        <Route path={USERS_ANALYTICS_PATH} element={<UsersAnalyticPage />} />
-        <Route path={ITEMS_ANALYTICS_PATH} element={<ItemAnalyticPage />} />
-        <Route path={APPS_ANALYTICS_PATH} element={<AppsAnalyticPage />} />
-        <Route path={EXPORT_ANALYTICS_PATH} element={<ExportAnalyticsPage />} />
+        <Route path={HOME_PATH} element={<HomePageWrapper />} />
+        <Route path={MY_ANALYTICS_PATH} element={<MyAnalyticsPage />} />
+
+        <Route
+          path={buildItemPath()}
+          element={
+            <Box id="navigatorContainer" width="100%">
+              <Navigator />
+              <ViewDataProvider>
+                <ItemPage />
+              </ViewDataProvider>
+            </Box>
+          }
+        >
+          <Route index element={<GeneralAnalyticsPage />} />
+          <Route path={USERS_ANALYTICS_PATH} element={<UsersAnalyticPage />} />
+          <Route path={ITEMS_ANALYTICS_PATH} element={<ItemAnalyticPage />} />
+          <Route path={APPS_ANALYTICS_PATH} element={<AppsAnalyticPage />} />
+          <Route
+            path={EXPORT_ANALYTICS_PATH}
+            element={<ExportAnalyticsPage />}
+          />
+        </Route>
       </Route>
     </Routes>
   );