+++ src/components/layout/PageWrapper.tsx
@@ -1,4 +1,4 @@
-import { Link, useLocation, useParams } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 
 import { Box, styled } from '@mui/material';
 
@@ -16,7 +16,6 @@ import { useAnalyticsTranslation } from '@/config/i18n';
 import { HOME_PATH } from '@/config/paths';
 
 import UserSwitchWrapper from '../common/UserSwitchWrapper';
-import Navigator from './Navigator';
 import Sidebar from './Sidebar';
 
 export const platformsHostsMap = defaultHostsMapper({
@@ -37,7 +36,6 @@ const LinkComponent = ({ children }: { children: JSX.Element }) => (
 const PageWrapper = ({ children }: { children: JSX.Element }): JSX.Element => {
   const { t } = useAnalyticsTranslation();
   const { itemId } = useParams();
-  const { pathname } = useLocation();
 
   const getNavigationEvents = usePlatformNavigation(platformsHostsMap, itemId);
 
@@ -52,7 +50,6 @@ const PageWrapper = ({ children }: { children: JSX.Element }): JSX.Element => {
       ...getNavigationEvents(Platform.Analytics),
     },
   };
-  const isRoot = pathname === HOME_PATH;
 
   return (
     <Main
@@ -72,11 +69,6 @@ const PageWrapper = ({ children }: { children: JSX.Element }): JSX.Element => {
       LinkComponent={LinkComponent}
     >
       <Box height="100%" display="flex" flexGrow={1} flexDirection="column">
-        {!isRoot && (
-          <Box id="navigatorContainer" width="100%">
-            <Navigator />
-          </Box>
-        )}
         {children}
       </Box>
     </Main>